fix tooltip and hover while dragging
